// src/protos/generated/sbai_system_alert_protos/sbai_system_alert_protos/aggregated_table.ts
export interface AggregatedTable {
  robotId: string;
  table: { [key: string]: Buffer };
}

export interface AggregatedTable_TableEntry {
  key: string;
  value: Buffer;
}

function createBaseAggregatedTable(): AggregatedTable {
  return { robotId: "", table: {} };
}

export const AggregatedTable: MessageFns<AggregatedTable> = {
  encode(message: AggregatedTable, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.robotId !== "") {
      writer.uint32(10).string(message.robotId);
    }
    Object.entries(message.table).forEach(([key, value]) => {
      AggregatedTable_TableEntry.encode({ key: key as any, value }, writer.uint32(18).fork()).join();
    });
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AggregatedTable {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAggregatedTable();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.robotId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          const entry2 = AggregatedTable_TableEntry.decode(reader, reader.uint32());
          if (entry2.value !== undefined) {
            message.table[entry2.key] = entry2.value;
          }
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AggregatedTable {
    return {
      robotId: isSet(object.robotId) ? globalThis.String(object.robotId) : "",
      table: isObject(object.table)
        ? Object.entries(object.table).reduce<{ [key: string]: Buffer }>((acc, [key, value]) => {
          acc[key] = Buffer.from(bytesFromBase64(value as string));
          return acc;
        }, {})
        : {},
    };
  },

  toJSON(message: AggregatedTable): unknown {
    const obj: any = {};
    if (message.robotId !== "") {
      obj.robotId = message.robotId;
    }
    if (message.table) {
      const entries = Object.entries(message.table);
      if (entries.length > 0) {
        obj.table = {};
        entries.forEach(([k, v]) => {
          obj.table[k] = base64FromBytes(v);
        });
      }
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AggregatedTable>, I>>(base?: I): AggregatedTable {
    return AggregatedTable.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AggregatedTable>, I>>(object: I): AggregatedTable {
    const message = createBaseAggregatedTable();
    message.robotId = object.robotId ?? "";
    message.table = Object.entries(object.table ?? {}).reduce<{ [key: string]: Buffer }>((acc, [key, value]) => {
      if (value !== undefined) {
        acc[key] = value;
      }
      return acc;
    }, {});
    return message;
  },
};

function createBaseAggregatedTable_TableEntry(): AggregatedTable_TableEntry {
  return { key: "", value: Buffer.alloc(0) };
}

export const AggregatedTable_TableEntry: MessageFns<AggregatedTable_TableEntry> = {
  encode(message: AggregatedTable_TableEntry, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.key !== "") {
      writer.uint32(10).string(message.key);
    }
    if (message.value.length !== 0) {
      writer.uint32(18).bytes(message.value);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AggregatedTable_TableEntry {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAggregatedTable_TableEntry();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.key = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.value = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AggregatedTable_TableEntry {
    return {
      key: isSet(object.key) ? globalThis.String(object.key) : "",
      value: isSet(object.value) ? Buffer.from(bytesFromBase64(object.value)) : Buffer.alloc(0),
    };
  },

  toJSON(message: AggregatedTable_TableEntry): unknown {
    const obj: any = {};
    if (message.key !== "") {
      obj.key = message.key;
    }
    if (message.value.length !== 0) {
      obj.value = base64FromBytes(message.value);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AggregatedTable_TableEntry>, I>>(base?: I): AggregatedTable_TableEntry {
    return AggregatedTable_TableEntry.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AggregatedTable_TableEntry>, I>>(object: I): AggregatedTable_TableEntry {
    const message = createBaseAggregatedTable_TableEntry();
    message.key = object.key ?? "";
    message.value = object.value ?? Buffer.alloc(0);
    return message;
  },
};

function bytesFromBase64(b64: string): Uint8Array {
  return Uint8Array.from(globalThis.Buffer.from(b64, "base64"));
}

function base64FromBytes(arr: Uint8Array): string {
  return globalThis.Buffer.from(arr).toString("base64");
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isObject(value: any): boolean {
  return typeof value === "object" && value !== null;
}

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}