// src/protos/generated/sbai_task_protos/sbai_task_protos/behavior_status.ts
export enum BehaviorState {
  BEHAVIOR_STATE_UNSPECIFIED = 0,
  BEHAVIOR_STATE_ACCEPTED = 1,
  BEHAVIOR_STATE_ACTIVE = 2,
  BEHAVIOR_STATE_COMPLETED = 3,
  BEHAVIOR_STATE_FAILED = 4,
  BEHAVIOR_STATE_CANCELED = 5,
  UNRECOGNIZED = -1,
}
// ...
export interface BehaviorStatus {
  behaviorState: BehaviorState;
  behaviorId: string;
}

function createBaseBehaviorStatus(): BehaviorStatus {
  return { behaviorState: 0, behaviorId: "" };
}

export const BehaviorStatus: MessageFns<BehaviorStatus> = {
  encode(message: BehaviorStatus, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.behaviorState !== 0) {
      writer.uint32(8).int32(message.behaviorState);
    }
    if (message.behaviorId !== "") {
      writer.uint32(18).string(message.behaviorId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BehaviorStatus {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBehaviorStatus();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.behaviorState = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.behaviorId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BehaviorStatus {
    return {
      behaviorState: isSet(object.behaviorState) ? behaviorStateFromJSON(object.behaviorState) : 0,
      behaviorId: isSet(object.behaviorId) ? globalThis.String(object.behaviorId) : "",
    };
  },

  toJSON(message: BehaviorStatus): unknown {
    const obj: any = {};
    if (message.behaviorState !== 0) {
      obj.behaviorState = behaviorStateToJSON(message.behaviorState);
    }
    if (message.behaviorId !== "") {
      obj.behaviorId = message.behaviorId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BehaviorStatus>, I>>(base?: I): BehaviorStatus {
    return BehaviorStatus.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BehaviorStatus>, I>>(object: I): BehaviorStatus {
    const message = createBaseBehaviorStatus();
    message.behaviorState = object.behaviorState ?? 0;
    message.behaviorId = object.behaviorId ?? "";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}